+++ test/resource.test.ts
@@ -1,0 +1,265 @@
+import { feathers } from '@feathersjs/feathers';
+import memory from 'feathers-memory';
+import { skip, tap, take } from 'rxjs/operators';
+import { beforeEach, describe, expect, it } from 'vitest';
+
+import { rx } from '../src';
+
+describe('reactive resources', () => {
+  let app, id, service;
+
+  describe('standard id', function () {
+    beforeEach(
+      () =>
+        new Promise<void>((done) => {
+          app = feathers()
+            .configure(rx({ idField: 'id' }))
+            .use(
+              '/messages',
+              memory({
+                multi: ['create']
+              })
+            );
+
+          service = app.service('messages');
+          service
+            .create({
+              text: 'A test message'
+            })
+            .then((message) => {
+              id = message.id;
+              done();
+            });
+        })
+    );
+    baseTests('id');
+  });
+
+  describe('custom id on service', function () {
+    beforeEach(
+      () =>
+        new Promise<void>((done) => {
+          app = feathers()
+            .configure(rx({ idField: 'id' }))
+            .use(
+              '/messages',
+              memory({
+                multi: ['create'],
+                id: 'customId'
+              })
+            );
+
+          service = app.service('messages').rx({ idField: 'customId' });
+
+          service
+            .create({
+              text: 'A test message'
+            })
+            .then((message) => {
+              id = message.customId;
+              done();
+            });
+        })
+    );
+    baseTests('customId');
+  });
+
+  describe('custom id on params', function () {
+    beforeEach(
+      () =>
+        new Promise<void>((done) => {
+          app = feathers()
+            .configure(rx({ idField: 'id' }))
+            .use(
+              '/messages',
+              memory({
+                multi: ['create'],
+                id: 'customId'
+              })
+            );
+
+          service = app.service('messages').hooks({
+            before: {
+              all: [
+                function (hook) {
+                  hook.params.rx = { idField: 'customId' };
+                }
+              ]
+            }
+          });
+
+          service
+            .create({
+              text: 'A test message'
+            })
+            .then((message) => {
+              id = message.customId;
+              done();
+            });
+        })
+    );
+    baseTests('customId');
+  });
+
+  function baseTests(customId) {
+    it('methods are still Promise compatible', () =>
+      new Promise<void>((done) => {
+        service.get(id).then((message) => {
+          expect(message).toStrictEqual({
+            [customId]: id,
+            text: 'A test message'
+          });
+          done();
+        }, done);
+      }));
+
+    it('.get as an observable', () =>
+      new Promise<void>((done) => {
+        service
+          .watch()
+          .get(id)
+          .pipe(take(1))
+          .subscribe((message) => {
+            expect(message).toStrictEqual({
+              [customId]: id,
+              text: 'A test message'
+            });
+            done();
+          }, done);
+      }));
+
+    it('lazy execution on subscription', () =>
+      new Promise<void>((done) => {
+        let ran = false;
+
+        app.use('/dummy', {
+          get(id) {
+            ran = true;
+
+            return Promise.resolve({
+              id,
+              description: `Do ${id}!`
+            });
+          }
+        });
+
+        const source = app.service('dummy').watch().get('dishes').pipe(take(1));
+
+        expect(ran).toStrictEqual(false);
+
+        source.subscribe((data) => {
+          expect(data).toStrictEqual({
+            id: 'dishes',
+            description: 'Do dishes!'
+          });
+          expect(ran).toStrictEqual(true);
+          done();
+        }, done);
+      }));
+
+    it('.update and .patch update existing stream', () =>
+      new Promise<void>((done) => {
+        const result = service.watch().get(id);
+
+        result.pipe(take(1)).subscribe((message) => {
+          expect(message).toStrictEqual({
+            [customId]: id,
+            text: 'A test message'
+          });
+          service.update(id, {
+            text: 'Updated',
+            prop: true
+          });
+        }, done);
+
+        result.pipe(skip(1), take(1)).subscribe((message) => {
+          expect(message).toStrictEqual({
+            [customId]: id,
+            text: 'Updated',
+            prop: true
+          });
+          service.patch(id, { text: 'Updated again' });
+        }, done);
+
+        result.pipe(skip(2), take(1)).subscribe((message) => {
+          expect(message).toStrictEqual({
+            [customId]: id,
+            text: 'Updated again',
+            prop: true
+          });
+          done();
+        }, done);
+      }));
+
+    it('.remove emits null', () =>
+      new Promise<void>((done) => {
+        service
+          .watch()
+          .get(id)
+          .pipe(take(2))
+          .subscribe((message) => {
+            if (message === null) {
+              done();
+            } else {
+              expect(message).toStrictEqual({
+                [customId]: id,
+                text: 'A test message'
+              });
+            }
+          }, done);
+
+        setTimeout(() => service.remove(id));
+      }));
+
+    it('injects options.pipe into observable chain (single operator)', () =>
+      new Promise<void>((done) => {
+        const options = {
+          pipe: tap(() => done())
+        };
+        service.watch(options).get(0).pipe(take(1)).subscribe();
+      }));
+
+    it('injects options.pipe into observable chain (array of operators)', () =>
+      new Promise<void>((done) => {
+        let i = 0;
+
+        const options = {
+          pipe: [
+            tap(() => i++),
+            tap(() => {
+              expect(i).toStrictEqual(1);
+              done();
+            })
+          ]
+        };
+        service.watch(options).get(0).pipe(take(1)).subscribe();
+      }));
+
+    it('.get uses caching', () =>
+      new Promise<void>((done) => {
+        const o1 = service.watch().get(0);
+        const o2 = service.watch().get(0);
+
+        expect(o2).toStrictEqual(o1);
+
+        done();
+      }));
+
+    it('clears cache after unsubscription', () =>
+      new Promise<void>((done) => {
+        const o1 = service.watch().get(0);
+        const o2 = service.watch().get(0);
+
+        const sub1 = o1.subscribe();
+        const sub2 = o2.subscribe();
+
+        sub1.unsubscribe();
+        sub2.unsubscribe();
+
+        expect(o2).toStrictEqual(o1);
+        expect(service.watch().get(0)).not.toStrictEqual(o1);
+
+        done();
+      }));
+  }
+});
